requirements.txt
# Requirements for the LinkedIn Scraping / Ice Breaker project
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-ollama>=0.2.3
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.2
webdriver-manager>=4.0.1
tweepy>=4.14.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
jinja2>=3.1.2
python-multipart>=0.0.6
ollama>=0.4.7 